# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SearchAnalysisResponse": {
        "description": "검색 분석 결과를 담는 응답 모델",
        "properties": {
          "entity_analysis": {
            "additionalProperties": true,
            "title": "Entity Analysis",
            "type": "object"
          },
          "execution_time_ms": {
            "title": "Execution Time Ms",
            "type": "number"
          },
          "performance_metrics": {
            "additionalProperties": true,
            "title": "Performance Metrics",
            "type": "object"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "results": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Results",
            "type": "array"
          },
          "search_strategy": {
            "additionalProperties": true,
            "title": "Search Strategy",
            "type": "object"
          }
        },
        "required": [
          "query",
          "execution_time_ms",
          "entity_analysis",
          "search_strategy",
          "results",
          "performance_metrics"
        ],
        "title": "SearchAnalysisResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "InternalApiKey": {
        "description": "AI 내부 호출용 키. Authorization Bearer 토큰을 사용할 수 있습니다.",
        "in": "header",
        "name": "X-Internal-Api-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "KBO AI Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/ai/search/": {
      "get": {
        "description": "검색 알고리즘을 자세히 분석하고 성능을 측정하는 디버깅 엔드포인트입니다.\n개발자가 검색 품질을 개선하기 위해 사용합니다.",
        "operationId": "debug_search_ai_search__get",
        "parameters": [
          {
            "description": "분석할 질문 또는 키워드",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "분석할 질문 또는 키워드",
              "minLength": 2,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "description": "검색 결과 개수",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "description": "검색 결과 개수",
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "시즌 연도 (예: 2025)",
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "시즌 연도 (예: 2025)",
              "title": "Year"
            }
          },
          {
            "description": "팀명 (예: LG, KIA)",
            "in": "query",
            "name": "team",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "팀명 (예: LG, KIA)",
              "title": "Team"
            }
          },
          {
            "description": "리그 코드 (예: KBO)",
            "in": "query",
            "name": "league",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "리그 코드 (예: KBO)",
              "title": "League"
            }
          },
          {
            "description": "다중 쿼리 검색 사용 여부",
            "in": "query",
            "name": "use_multi_query",
            "required": false,
            "schema": {
              "default": true,
              "description": "다중 쿼리 검색 사용 여부",
              "title": "Use Multi Query",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "X-Internal-Api-Key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Internal-Api-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchAnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "InternalApiKey": []
          }
        ],
        "summary": "Debug Search",
        "tags": [
          "search"
        ]
      }
    },
    "/ai/search/compare-methods": {
      "get": {
        "description": "단일 쿼리 검색과 다중 쿼리 검색 성능을 비교하는 엔드포인트입니다.",
        "operationId": "compare_search_methods_ai_search_compare_methods_get",
        "parameters": [
          {
            "description": "비교 테스트할 질문",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "비교 테스트할 질문",
              "minLength": 2,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Internal-Api-Key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Internal-Api-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "InternalApiKey": []
          }
        ],
        "summary": "Compare Search Methods",
        "tags": [
          "search"
        ]
      }
    },
    "/ai/search/test-entity-extraction": {
      "get": {
        "description": "질문에서 엔티티 추출이 올바르게 작동하는지 테스트하는 엔드포인트입니다.",
        "operationId": "test_entity_extraction_ai_search_test_entity_extraction_get",
        "parameters": [
          {
            "description": "엔티티 추출을 테스트할 질문",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "엔티티 추출을 테스트할 질문",
              "minLength": 2,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Internal-Api-Key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Internal-Api-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "InternalApiKey": []
          }
        ],
        "summary": "Test Entity Extraction",
        "tags": [
          "search"
        ]
      }
    },
...
  }
}